Rename and move GraphCentralityUtilities -> ClosenessCentralityUtilities.
